Tune VDR decryption batch sizes (#11478)
## Describe your changes
Previously, we were decrypting up to 2000 fields or 2000 documents simultaneously. This actually worked okay for fields, but for documents it was clear from the traces that this was way too many simultaneous connections - we saw a latency waterfall that suggested we were over oversubscribing the network or CPU. In this PR, we reduce the batch size to try to improve the overall performance.

## Caveats

## Test Plan