Fix bug in vault DR batch query (#11481)
## Describe your changes
The new test describes the bug in the query. If a seqno isn't completely handled by a single batch, some DLs for that seqno won't every be written. We'll have to rewrite this query entirely to accommodate manifests, so we can fix any impact of the bug in that backfill.

## Caveats

## Test Plan