React obc inline auth 1 (#11498)
## Describe your changes
We have been wrongly calling `vaultingToken` to be `authToken` so far on
the React OBC sdk. This PR corrects it. Now it's aligned with our RN and
Flutter SDKs.

Note: we still accept `authToken` to be passed as prop to the
`Provider`. That is because, we will use that token to identify user in
another PR soon.


## Caveats


## Test Plan