Remove warning about no composite fingerprints when backfilling (#11574)
## Describe your changes
In a previous PR, I exposed an endpoint to backfill composite
fingerprints in batches. This is working as expected, until it runs into
a user without any composite fingerprints, at which point, a 500 will be
returned.

This PR stops the endpoint from throwing an error when no composite
fingerprints should be/were generated.

## Caveats
N/A

## Test Plan
Run composite fingerprint backfill, ensure no errors are thrown for this
reason.