Implement authetication using auth token (#11406)
## Describe your changes
Tenants can make a request `requiresAuth` which with returns an object
we values:

```tsx
{
	requiresAuth: boolean,
	method: "authToken" | "emailAndPhone"
}
```

**For method authToken**

We will use the method authToken if the tenant passes a `valid`
`authToken` . `authToken` is passed through the `Provider` as a prop.
The first time `requiresAuth` is called, we will run an identify query
with the `authToken` . Based on the response, we will decide a status of
the `authToken` which we will also keep in the internal provider. The
status’ type is as follows:

```tsx
authTokenStatus?: "validWithSufficientScope" | "validWithInsufficientScope" | "invalid" 
```

If the status is:

- `validWithSufficientScope` - we don’t need to reauth
- `validWithInsufficientScope` - we need to reauth with method being
`authToken`
- `invalid`  - we need to reauth with method being `emailAndPhone`

If the tenant calls `requiresAuth` again, we will simply use the
existing result from the previous query. We can do this because, within
one provider, there can only be one `authToken` and it can’t be updated
after the first mount.

**For method emailAndPhone**

If the tenant didn’t provide any `authToken` or if the `authToken`
provided is `invalid` we will reauth with `emailAndPhone`

Additionally, we add/update a few more functions:

Instead of providing one function to tenants for creating challenges, we
now provide two functions `createEmailPhoneBasedChallenge({String email,
String phone})` , `createAuthTokenBasedChallenge()` . Note that tenant
must provide both email and phone to `createEmailPhoneBasedChallenge` -
we don’t support email-only for now to reduce complications. However, we
do support email only login, if the user went through an email-only
signup before via some other SDK or our web flow.

*NOTE if the tenant attempt to auth using a method not returned by
`requiresAuth` , we will throw an error. However, we will allow user to
reauth as many times as they want as long as they use the correct auth
method.*

### Additionally
this PR cleans up a bit and addresses some performance issues.

### With auth token

https://github.com/user-attachments/assets/6cb8e106-fcbf-4f81-903c-d02f1a783360

### Without auth token

https://github.com/user-attachments/assets/192743e1-dd5e-44d4-a1ac-4f913bf5f2ce

### Tested cases
- With `null` auth token
- wrong auth token
- empty auth token
- valid auth token
- trying wrong auth method






## Caveats


## Test Plan