Don't wait for selfie when we repeatedly receive SelfieTooDark (#11646)
## Describe your changes

https://onefootprint.slack.com/archives/C06SZL2F53Q/p1725511130597529?thread_ts=1725501678.734719&cid=C06SZL2F53Q

There are a few cases from incode where we received the `SelfieTooDark`
error on repeated selfie attempts. The incode machine was getting stuck
and leaving these users in a `pending` state.

We have some existing logic that stops waiting for selfie (and moves the
user into Failed) if we get repeated `SelfieFaceNotFound` - it feels
like we should also stop waiting for selfie if we get repeated
`SelfieTooDark` errors.


## Caveats
I'm going to at least merge this to get the two users un-stuck from
pending. But I don't have a ton of confidence yet to make this change
wholly, so i might revert it after getting these users unstuck knowing
that Dave will be back next week to opine 😛


## Test Plan