Spawn tokio tasks to encrypt and write VDR blobs (#11696)
## Describe your changes
Previously, we weren't spawning tasks for the async IO part of a VDR
batch (writing to S3). I think this meant that this IO work was getting
serialized on the same thread (i.e. concurrency but no parallelism).
This would explain why the throughput we were getting writing to S3 is
close to my single-threaded benchmarks, and increasing the concurrency
actually decreased throughput.

## Caveats


## Test Plan