[VDR Performance] Lazily spawn Tokio tasks for writing blobs (#11760)
## Describe your changes
I was confused why the minimum latency for VDR blob writes (~500 byte S3
objects) was in the hundreds of milliseconds. After doing some test
benchmarks, I found a subtle bug in the parallel processing code.

We were collecting a `Vec` of `JoinHandle`s from `tokio::task::spawn`,
which meant that the entire batch of tasks was spawned immediately, and
the concurrency limit was only applicable to polling those join handles.
This meant we effectively had no concurrency limit -- we were processing
the entire batch simultaneously.

By constructing a lazy iterator of `JoinHandle`s and not collecting
them, the Tokio task concurrency is controlled by the `buffer_unordered`
stream consumer.

I had previously reduced the concurrency setting since I though it was
set too high, but was actually observing the bug described above. So now
we can raise it again. Some benchmarking on a 2 core machine showed that
a concurrency of 128 is approximately the right concurrency for optimal
end-to-end latency for a batch of 1000.

## Caveats


## Test Plan