Updated composite fingerprint backfill script to include DI's that contain an alias (#11772)
## Describe your changes
Our current composite fingerprint backfill script does not work for fingerprints that contain an alias. This PR exposes the route for use, and enumerates all DI's in a scoped vault to generate the necessary fingerprints.

## Caveats
N/A

## Test Plan
Test backfilling composite fingerprints locally on a user with bank accounts and credit cards to ensure that all composite fingerprints are present.