Remove exposed handler for backfilling composite fingerprints (#11788)
## Describe your changes
In a previous PR, we exposed a route for backfilling composite fingerprints. This PR removes the route to prevent it from being accidentally invoked.

## Caveats
This route will need to be re-exposed when we want to run another backfill.

## Test Plan
Send a request to `/private/backfill/batch_composite_fingerprints`, and make sure that it returns a 404.