Fix composer citizenship editing bug (#11741)
## Describe your changes

This is, in fact, a bug! I ran this locally and found
that`usFormLegalStatus` was showing up as undefined when we changed
citizenship. This would cause the bug — the validation logic would have
no idea that the status was "Green card", because `usFormLegalStatus`
will be undefined unless the value is changed.

This is because `usFormLegalStatus` is based on a `useWatch`; here, I
add a `getValues` call so that there is a fallback. I tested this and it
solves the problem.

## Caveats

N/A

## Test Plan

- tested using the composer user that failed, error goes away
- tested changing the formationCountry and residenceCountry and the
state inputs changed correctly