Restructure cmdk component in prep for edit redesign (#11940)
## Describe your changes
Modify the dialog open/close logic in `cmdk.tsx` to be like
`actions.tsx`. This is in preparation for the addition of
`EditVaultDrawer`, which will need to be opened differently than the
edit view is right now. **Should not change any ui/functionality — only
the component's internal logic!**

Specifically, now instead of having one state for each dialog to store
whether that dialog is open/closed (eg `retriggerDialogOpen` +
`editVaultDialogOpen` + `manualReviewDialogOpen`...), we have 1 state
`openDialog` which stores an `enum` of what is open (`editVault` or
`reviewPass` etc, or `null` if no dialog is open)

## Caveats


## Test Plan
If you could test a bit locally/using the deployment that would be
great. I also tested, but haven't used this feature a ton so am less
familiar with expected behavior.