[BE-1553] Simplify DataLifetime source specification (#11872)
## Describe your changes
Some of the write validation logic needs pretty granular information about the callsite of the vault write - sometimes a DI can only be set by a few callsites. Previously, we had a web of a mess of DataLifetimeSource, actor, and DataRequestSource passed around to determine this. This consolidates all three of those args into one much more explicit argument

## Caveats
Still is not "simple," but is simpler ¯\_(ツ)_/¯

## Test Plan