[BE-1562] Allow tenant to write id.phone_number and id.email (#11901)
## Describe your changes
For tenants opted into a feature flag, allows these tenants to write `id.phone_number` and `id.email` _even_ if there are existing verified values.

In order to preserve the user's existing login behavior, this will copy over the "verification status" of the user's old contact info into the new contact info. So, if the user's existing phone number was verified, the new phone number will also be marked as OTP verified. This is a pretty arbitrary implicit decision, but this felt simpler than explaining to tenants the difference between, say, an  `id.verified_phone_number` and `id.phone_number` since the differences are so so subtle.

## Caveats
This starts to ignore the previously in-progress work to add `id.verified_phone_number` and `id.verified_email`

## Test Plan