[BE-520] Remove validation on dangling identifiers (#11959)
## Describe your changes
https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1726486113366059

This validation of "dangling" identifiers was initially added to make it
hard to accidentally update only part of CDO. For example, it would
error if you updated `id.address_line1` without updating the rest of the
components of the address, since you might then be writing an
inconsistent address.

This has required lots of really complex logic on our clients to uphold
this invariant that we send the full CDOs at the same time. We relaxed
this constraint on tenant-facing vaulting APIs, and now we're finally
relaxing the constraint on bifrost-facing APIs


Now, there's probably some cleanup we can do on our clients to not send
all fields of the CDO when only one is updated. [For ex](
https://github.com/onefootprint/monorepo/blob/f26ac0ab51639a2486e30ab7ea43a427a9b1b46e/frontend/packages/idv/src/plugins/collect-kyc-data/hooks/use-sync-data/utils/get-request-data/get-request-data.ts#L50-L52)

## Caveats


## Test Plan