[BE-1536] Explicitly mark only the exact contact info as verified (#11936)
## Describe your changes
Gets rid of any bugs where the piece of data to which the challenge was initiated has changed by the time the challenge was verified - now, we pass the exact lifetime ID in the challenge token.

Much easier to do now that signup and login challenges share so much code!

## Caveats

## Test Plan