Make VDR blob writes idempotent (#12003)
## Describe your changes
This two-part change simplifies VDR correctness analysis by allowing us to model both S3 blobs and the `vault_dr_blob` table as sets of DLs, rather than as a multiset/bag. It enforces that we don't process & commit DLs multiple times for a given config. First, we make S3 writes idempotent so the key for each DL is deterministic. Then in the next PR we enforce that we don't write duplicates to the DB with a new uniqueness constraint on `vault_dr_blob(config_id, data_lifetime_id)`. With a determistic key, this is also enforced by the existing uniqueness constraint on the `vault_dr_blob(bucket_path)`, but the new index will enforce the constraint regardless of the S3 key structure (i.e. including previously generated keys).

## Caveats

## Test Plan