chore(test-utils): [FE-629] react-query v5 migration (#11887)
## Describe your changes

Title! Per the migration guide, [custom logging was deprected and is now
removed](https://tanstack.com/query/latest/docs/framework/react/guides/migrating-to-v5#the-deprecated-custom-logger-has-been-removed)
so we had to get rid of it. Change here so that we just copy over the
test-utils. We will remove this eventually - I picked
`updated-utils-testing` so I could more easily search for it, versus if
it was `test-utils-new` (where, when searching for `test-utils`, I'd
find the new one, too)

## Caveats

N/A

## Test Plan

CI