Rename `obtok_` to `pbtok_ (#12012)
## Describe your changes
Chatted with rafa: we're going to start accepting the onboarding session
tokens in the same SDK argument as the playbook publishable key. this
argument will be called `playbookKey`.

So i think it makes more sense for these tokens to look like `pbtok_xxx`
to mirror the `pb_live_xxxx` keys

## Caveats


## Test Plan
We had some explicit matching on the prefix of this token before - I've
migrated that client code to also expect `pbtok_` values now too. I
tested running bifrost locally and invoking it with a legacy `obtok_`
value and a modern `pbtok_` value - both worked properly

The e2e tests here are going to fail because they run against the
development version of the backend, not the version of the backend from
this PR