Remove `obConfig` from GET /org/sdk_args (#12030)
## Describe your changes
Cannot merge until after #12027 and #12028 are merged.

There are becoming so many ways for the playbook to be specific via SDK
args that we won't be able to reliably extract it:
1. playbook key
2.  user auth token, whose session has a playbook key specified
3. onboarding session token, whose session has a playbook key specific

## Caveats
Unfortunately, we lose a little bit of visibility into on behalf of
which tenants each sdk_args API request is made since we're no longer
doing the work to extract the playbook. But this information was already
was misleadingly empty for SDK args tokens made via (2) or (3) - it
might be less damaging to just not have this information at all than to
have it only partially available

We also lose validation that the playbook key passed into SDK args is
valid. But perhaps that was an unnecessary validation to begin with -
SDK args is mostly meant to be a JSON blob storage to pass arguments
from outside the SDK to inside

## Test Plan