Footprint-react: process (#11973)
## Describe your changes

Context: some customers don't want to use the handoff flow, given they
don't want to collect passkey and documents. This is the case of
composer. So they asked to see if we could implement something to
process the flow, without opening the handoff.

This PR exposes a process method, which will finalize the flow and
return the validation token if everything is all right. In case the flow
is still pending a requirement (because you haven't collected or you got
a step-up), we'll return an error, and then, you'll need to fallback to
handoff:

```typescript
const handleSubmit = async (formValues: FormValues) => {
    setBusy(true);
    try {
      await fp.vault(formValues);
      const { validationToken } = await fp.process();
      onDone(validationToken);
    } catch (e) {
      if (e instanceof InlineProcessError) {
        fp.handoff({ onComplete: onDone });
      }
    } finally {
      setBusy(false);
    }
  };
```

Demo:


https://github.com/user-attachments/assets/b405f7d4-84bf-4fe0-86e7-3de2af9e33b1



## Caveats


## Test Plan