Require a lock when advancing seqnos (#12064)
## Describe your changes
While working on VDR correctness, I realized a nuance of our seqno usage with DataLifetimes that affects how we can interpret them as an ordering.

Postgres sequences are an exception to read committed isolation -- `nextval` is never rolled back.

Therefore, we can't get too comfortable with using seqnos for ordering without careful analysis. Seqnos are only a valid ordering of events within the scope of a single locked resource in a transaction. For example, means they can not be interpreted as an ordering across scoped vaults.

Additionally, it means that incrementing seqnos must be done only while holding a lock on the relevant resource. Consider the following buggy case of two transactions writing DLs to the same vault:
1. `Txn_A` fetches the next seqno: `n.
2. `Txn_B` fetches the next seqno: `n+1.`
3. `Txn_B `takes a lock on `Vault_1`, reads some data, and based on that data writes a new `DL_B` at seqno `n+1` and commits.
4. `Txn_A `takes a lock on Vault_1, reads some data, and based on that data write a new `DL_A` at `n` and commits.
5. Client reads back `DL_A` as being written before `DL_B`

Because the seqnos are advanced without holding a lock, the writes will appear to be out of order when read back in order of seqno.

The start of the solution here is to require a `ScopedVault` lock when we advance the `DataLifetimeSeqno`. We'll make this stronger in the following PRs.

## Caveats

## Test Plan