Improve safety of vault wrapper data validity by moving WriteableVw internally (#12065)
## Describe your changes
This is really a consequence of the changes made in following PR due to some lifetime issues, but it's a good change anyway. I have to order it here since it's required to get the following changes to compile.

There's future work of this kind as well. Currently, writing documents appears to be "unsafe" in that vault wrappers are invalidated without consuming `self`.

## Caveats

## Test Plan