Footprint-provider: add isReady (#12107)
## Describe your changes

The composer team had a situation where they created the challenge right
after the provider appeared because they have already user phone and
email. However, the provider needs to fetch the onboarding config first.
Because the challenge was created very fast, without collecting the data
from an input, the request for fetching the ob config wasn't finalized
yet, so they were getting an error when creating the challenge

This PR introduces a flag `isReady`, so customers will be able to check.
This was inspired by next router isReady, and it was the simplest option
we found for now. In the future, we might also integrate with react
suspense.


## Caveats


## Test Plan