[BE-1580] Write attributes of log lines directly without `fields.` prefix (#12105)
## Describe your changes
We chose to arbitrarily add a `fields.` prefix to all attributes specified on a log line when exporting to datadog to ensure that the user-provided attribute names wouldn't conflict with the reserved fields set automatically by our infra.

This adds a wrapper around the attribute serialization code that will automatically add a `CONFLICTING_FIELD.` prefix to any field names that have already been set, allowing us to safely serialize user-provided fields _after_ the reserved fields when we are exporting a log line to datadog.

This will make it a bit easier to write queries - if you see an attribute in code, you can filter directly on that attribute name!

## Caveats

## Test Plan
From this code:
```
    tracing::info!(
        tenant_id = "flerp1",
        tenant_id = "flerp2",
        vault_id = "not specified so i can write it",
        "my log line"
    );
```

This log line was emitted:
```
{
  "timestamp": "2024-09-24T06:15:08.006121Z",
  "level": "INFO",
  "target": "api_route_onboarding::session::index",
  "filename": "components/api/routes/onboarding/src/session/index.rs",
  "line_number": 71,
  "http.method": "POST",
  "http.route": "/onboarding/session",
  "http.host": "localhost:8000",
  "http.client_ip": "127.0.0.1",
  "http.user_agent": "curl/8.7.1",
  "http.target": "/onboarding/session",
  "request_id": "24bfaaa8-121d-44c6-bbe0-ab271a8c2daf",
  "server_git_hash": "a0ca462eeac764e8074bbd394416e69611d45507",
  "is_integration_test_req": false,
  "route": "POST /onboarding/session",
  "ip_address": "127.0.0.1",
  "tenant_id": "_private_it_org_2",
  "is_live": false,
  "auth_method": "secret_key",
  "dd.trace_id": "0b00a02cdf5554bbba92f25d38059552",
  "dd.span_id": "e482c8430996eb19",
  "trace_id": "0b00a02cdf5554bbba92f25d38059552",
  "fields": {
    "message": "my log line",
    "tenant_id": "flerp2",
    "vault_id": "not specified so i can write it"
  },
  "message": "my log line",
  "CONFLICTING_FIELD.tenant_id": "flerp1",
  "CONFLICTING_FIELD.CONFLICTING_FIELD.tenant_id": "flerp2",
  "vault_id": "not specified so i can write it"
}
```