[PR-163] getBankAccounts (#12155)
## Describe your changes

In our card logic, we use a helper called `getCards` to get all of the
card DIs. This PR puts the `getCards` logic into its own folder (it was
just in `utils` before), and adds the helper + tests for
`getBankAccounts`. We now also make changes to make both more efficient.
We will use this in the ultimate new card fieldset. Also a handful of
additions re typing.

## Caveats

N/A 

## Test Plan

`get-bank-accounts.test.ts`, `get-cards.test.ts` and pre-existing tests.
This is not used anywhere so it's quite low risk