[BE-1571] Support `allow_reonboard` flag in onboarding session token (#12035)
## Describe your changes
Supports creating an onboarding session token with an `allow_reonboard` flag. This is saved in the onboarding session token, which is later passed to `POST /hosted/identify` where we read back the value of `allow_reonboard`

## Caveats
There's a little bit of risk that the client could change to not call `POST /hosted/identify` first and thus won't create a token with the `allow_reonboard` value properly propagated. We should probably write an e2e test for that case... but it could be hard to uphold across all of our SDKs, especially as we start reimplementing the identify flow in many different places

Still need to make the frontend change to accept onboarding session tokens as the `publicKey` argument in the SDK for this to fully work e2e

## Test Plan
Integration tests