[BE-1578][1/2] Support passing onboarding session token as `publicKey` SDK argument (#12054)
## Describe your changes
In #12021, the backend started to accept onboarding session tokens (`pbtok_`) in every place that the playbook publishable key can be accepted. Now, onboarding session tokens can be passed into the sdk as the `publicKey` argument, and bifrost will magically treat this value as a replacement of the playbook's public key.

## Caveats
Yieldstreet is invoking the SDK with an onboarding session token in the `authToken` argument. I kept one e2e test using this legacy yieldstreet flow. We can get them to migrate off of `authToken` and onto the new `publicKey` argument. There's some logic in this PR too to extract an ob session token from the `authToken` argument for backwards compatibility

We were talking about renaming `publicKey` -> `playbookKey`. If we do this, we can ask them to migrate to use `playbookKey` instead of `authToken` in one go.

## Test Plan
Migrated end-to-end tests to pass the ob session token as the `publicKey`