Add struct that represents a request made to sentilink (#12040)
## Describe your changes
This PR adds an enum that represents a request to Sentilink. The idea
here is that we don't have to worry about handling `Result`/errors, just
need to handle the enumerated possible cases. We implement the logic
that maps a raw request into this enum as well

The only errors that aren't encapsulated here are ones at the actual
`reqwest` level

## Caveats


## Test Plan
wrote unit tests