Only create validation token after process success (#12336)
## Describe your changes

We were generating a validation token at the same time of getting the
onboarding status. However, once we do onboarding status, we may need to
do a doc step up. In this case, we should not generate a validation
token

This PR will now just create a validation token if process/onboarding
status are all success.


## Caveats


## Test Plan