Refactor fingerprint generation logic in FingeprintedDataRequest (#12314)
## Describe your changes
Right now, the logic for generating fingerprints inside of FingerprintedDataRequest is quite complicated.  This PR is an attempt to clean up this logic, and make what we are doing easily greppable.

NOTE: I need to look at this with fresh eyes to make sure that I didn't leave anything out.

## Caveats
N/A

## Test Plan
* Ensure correct fingerprints are being generated, and that duplicate fingerprints are not being generated.