Refactor Lexis match logic (#12320)
## Describe your changes
refactor our lexis matching logic to be more generic so we can add additional logic to test impact

## Caveats
we could do this in a notebook too, but this might be more straightforward to log a bit. before we put something into prod we can recreate the logic in a jupyter notebook and backtest on sample records. but it requires doing a bunch of json parsing which is amenable to bugs as well... so this seemed the fastest to get logs

## Test Plan