[PR-223] separate bank DI ordering (#12522)
## Describe your changes

Previously, uncaught because we were doing a sort, we weren't actually
handling sorting for bank account DIs. This was also obscured because
one of the unit tests didn't present the attributes as out of order.
Made this a bit more scalable by splitting the sorting functions, fixed
the test, and added in translations - this will allow us to remove
sorting on the upstream PR.

I checked with Pedro, and this is the order we want! 

## Caveats

N/A

## Test Plan

`get-dis.test.ts`