[BE-1592] Backfill `business.kyced_benficial_owners` (#12167)
## Describe your changes
Since we started requiring `business_kyced_beneficial_owners` instead of
`business_beneficial_owners` in all playbooks in #12470, this backfills
all instances of business vaults that have `business.beneficial_owners`
into `business.kyced_beneficial_owners`.



## Caveats
Here, we only backfill vaults that have a _single_ BO - the logic is a
little complex when a business has two BOs, and it would make the
business un-reonboardable since the secondary BOs wouldn't have a
phone/email associated. For all business vaults that _aren't_ backfilled
here, if they reonboard they will simply be asked to provide the BOs
again

We also omit adding any entries to the `business_owner` table, a luxury
we can only have if we are only migrated vaults that have one business
owner. If the vault was made via bifrost, it will have the primary BO
already in the `business_owner` table. If it was made via API, it will
have no rows in the `business_owner` table


## Test Plan