Add DataIdentifier for `business.beneficial_owners.xxx.yyy` (#12163)
## Describe your changes
Instead of our weird JSON data structure, we're going to start saving
each piece of each beneficial owner's data in a separate DI under the
business. The new DIs are serialized like
`business.beneficial_owners.{link_id}.{data identifier for user}`. In
this newer model, we also require that there is a `beneficial_owner` row
in the database (and thus a link ID) for every owner of the business
whose data we want to vault.

This newer model also nests IDKs, so in the future we can even support
bootstrapping in _any_ identity data for the beneficial owner

We're going to start vaulting these new DIs any time
`business.kyced_beneficial_owners `is written. In the future, we'll have
a separate set of REST-y API for both bifrsot _and_ the tenant to
`GET/POST/DELETE/PATCH /businesses/<>/owners`

## Caveats


## Test Plan