Add case of select custom in dialogs (#12552)
Because of how the content is positioned, the current `SelectCustom`
component fails inside dialogs. To simplify this behavior, we can use
the 'popper' position and always take the width of the trigger. Added
that story to storybooks and fixed props so we can use the available
`var(--radix-select-trigger-width)` for the width.