Update footprint-dr CLI to more gracefully handle partially written vault data (#12460)
## Describe your changes
If blobs are written for an fp_id but no manifests have committed those blobs, the fp_id will show up in the efficient S3 fp_id list procedure. To keep `list-vaults` fast (not requiring probing for a manifest), we're leaving the fp_id list procedure as is. However, we now skip over these in-progress fp_ids in the `list-records` call.

## Caveats

## Test Plan