Fix bug that would cause empty VDR manifests to not be written (#12461)
## Describe your changes
The TLA+ model stage `GetDlsForCompleteVaultVersionBatch` was correct, but the SQL/Rust translation omitted SVVs with no active DLs. I found this by testing on an integration test tenant in ephemeral, but it's a simple enough example that it should have be caught in integration tests. Better tests are a WIP...

Thinking about this raised the question of whether we should back up an empty "manifest.0.json" for all vaults. Currently, if there's no data ever written to a vault, the fp_id is completely absent from VDR. This is somewhat simple to change in the future if we reconsider.

## Caveats

## Test Plan