[PR-237] select-all refactor (#12583)
## Describe your changes

Because we have the "Select All" component at the top level of `financial-fieldset` , but we no longer control the DIs at that level, we needed some way to make them accessible.

My sense is it was going to be much more complex to deal with all of this "Select all" logic within each, so I figured we could just update the DIs globally whenever we change from card to card, bank account to bank account, or between them. Between them can only take place while not decrypting because we hide the segmented control during decryption, which covers that case.

However, what I also learned, which is great, is that this really intuitively handles the case where we swap between cards or bank accounts while decrypting. I checked with Pedro, and we actually don't want to disable the `card-selector` or `bank-account-selector` in that case.

NOTE: I also remove the `useMemo` logic here because we deemed it unecessary.
<img src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/3e5e181e-29ed-44e0-9464-faf71521c0f6.png" width="400" />

## Caveats

## Test Plan

Tested this a bunch in UI, but I'd welcome more! I'm using `fp_id_test_veBjtQOzj0ulQbdD30cQqK` on dev, so should work locally or on a preview link.