[PR-231] shared duplicate user component (#12542)
## Describe your changes

Title! This used to be called `duplicate-card` but it is actually showing duplicate users, so we can share it between the card and bank account duplicates. This is just UI and, because this is in a drawer that remains hidden, it doesn't show to users yet.

<img width="400" alt="Screenshot 2024-10-08 at 11 31 14 AM" src="https://github.com/user-attachments/assets/1cacd4c9-c715-43f4-bb5e-86b6b644ff64">

## Caveats

Right now this essentially shows all of the duplicates indiscriminately - in the future, this will show only the duplicates for one card. This doesn't show to users right now! It's just code we can merge in already / is meaningful UI. There's also more UI that needs to come for the card and bank account duplicate drawers.

## Test Plan

Four tests in `duplicate-user.test.tsx`. In UI as well, and Pedro reviewed / approved.

<img src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/2ebdef57-e767-4710-9295-fc8cf57a06fa.png" alt="Screenshot" width="300"/>

For testing locally, you can use: [this user](http://localhost:3001/users/fp_id_4omYsf3bK01pBjKtVy4a0M), for the preview environment you can use [this user](https://dashboard-git-lucas-pr-154.preview.onefootprint.com/users/fp_id_4omYsf3bK01pBjKtVy4a0M).