[PR-233] shared dupes loading state (#12554)
## Describe your changes

Title! In the case that dupes are still loading, which is pretty unlikely in this case because there will be time to load before the user clicks on this, we show the shimmer. This is only currently implemented in the card drawer, but because it is mocking the same `duplicate-user` as bank accounts will, we can use it for both (hence why the component is higher up).

Checked with Pedro and he gave me this — we'll always show two "users" loading because we won't have the exact number before we show.

https://github.com/user-attachments/assets/ae57d904-0732-47ee-9d6d-0a9d21a0b3dc

## Caveats

The current implementation of the shimmers assumes some text that is not actually there currently. It's coming in the next PR.

## Test Plan

`card-duplicate-drawer.test.tsx`  - also Pedro says this looks good!